# api-gateway/swagger.yaml
openapi: 3.0.3
info:
  title: CommerceAI Pro API
  description: |
    API Gateway pour la plateforme CommerceAI Pro - Intelligence Artificielle pour E-commerce
    
    Cette API fournit un accès complet aux fonctionnalités de la plateforme :
    - Authentification et gestion des utilisateurs
    - Gestion des organisations et rôles
    - Support client et chat en temps réel
    - Analytics et métriques
    - Intégrations avec les plateformes e-commerce
    - Agents IA spécialisés
    
    ## Authentification
    
    L'API utilise l'authentification JWT. Incluez le token dans l'en-tête Authorization :
    ```
    Authorization: Bearer <votre-token-jwt>
    ```
    
    ## Codes de statut
    
    - `200` - Succès
    - `201` - Créé avec succès
    - `400` - Erreur de validation
    - `401` - Non authentifié
    - `403` - Accès refusé
    - `404` - Ressource non trouvée
    - `429` - Limite de taux dépassée
    - `500` - Erreur serveur
    
  version: 1.0.0
  contact:
    name: Support CommerceAI Pro
    email: support@commerceai-pro.com
    url: https://commerceai-pro.com
  license:
    name: Propriétaire
    url: https://commerceai-pro.com/license
servers:
  - url: https://api.commerceai-pro.com
    description: Production
  - url: https://staging-api.commerceai-pro.com
    description: Staging
  - url: http://localhost:4000
    description: Développement local
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Erreur lors de l'opération"
        error:
          type: string
          example: "Détails de l'erreur"
        timestamp:
          type: string
          format: date-time
    Success:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Opération réussie"
        data:
          type: object
        timestamp:
          type: string
          format: date-time
    User:
      type: object
      properties:
        id:
          type: string
          example: "user_123"
        email:
          type: string
          format: email
          example: "user@example.com"
        firstName:
          type: string
          example: "John"
        lastName:
          type: string
          example: "Doe"
        role:
          type: string
          enum: [user, admin, support]
          example: "user"
        organizationId:
          type: string
          example: "org_123"
        isEmailVerified:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
...
security:
  - BearerAuth: []
paths:
...
  /api/auth/register:
    post:
      tags:
        - Authentification
      summary: Inscription d'un nouvel utilisateur
      description: Créer un nouveau compte utilisateur
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - firstName
                - lastName
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  minLength: 8
                  example: "motdepasse123"
                firstName:
                  type: string
                  example: "John"
                lastName:
                  type: string
                  example: "Doe"
                organizationName:
                  type: string
                  example: "Mon Entreprise"
      responses:
        '201':
          description: Utilisateur créé avec succès
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '400':
          description: Données invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email déjà utilisé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/login:
    post:
      tags:
        - Authentification
      summary: Connexion utilisateur
      description: Authentifier un utilisateur et obtenir un token JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  example: "motdepasse123"
                rememberMe:
                  type: boolean
                  example: false
      responses:
        '200':
          description: Connexion réussie
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          token:
                            type: string
                            example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                          user:
                            $ref: '#/components/schemas/User'
        '401':
          description: Identifiants invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/auth/me:
    get:
      tags:
        - Authentification
      summary: Profil utilisateur actuel
      description: Récupérer les informations de l'utilisateur connecté
      responses:
        '200':
          description: Profil utilisateur
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  # Support
...